refactor(ic-certified-assets): extract asset.rs from state_machine
Pull Asset, AssetEncoding, EncodedAsset, AssetDetails, AssetEncodingDetails,
Timestamp, and the asset-domain helpers (build_headers, on_asset_change,
delete_preexisting_asset_hashes, insert_new_response_hashes_for_encoding,
aliases_of, aliased_by, is_html_key, encoding_certification_order, plus
DEFAULT_ALIAS_ENABLED and STATUS_CODES_TO_CERTIFY) out of the 1500-line
state_machine/mod.rs into a dedicated top-level asset.rs.

These items form a coherent layer below State: the certification routine,
the per-encoding hash computation, the response renderer, and the key/HTML
heuristics. Keeping them in state_machine made it hard to see where one
concern ended and the next began. After the move state_machine/mod.rs is
~270 lines lighter and focuses on State + commit_batch.

Public surface change: Asset, AssetEncoding, EncodedAsset, AssetDetails,
AssetEncodingDetails, encoding_certification_order, and Timestamp now live
under `ic_certified_assets::asset` instead of `::state_machine`. The
`export_canister_methods!` macro and the in-tree `state_machine/stable.rs`
have been updated to follow the move. No external callers in this workspace
reference these symbols by path, so no downstream changes are needed.

`Asset::allow_raw_access` was bumped from module-private to `pub`; it is
called by `State::build_http_response` which now lives across a module
boundary.
